Refactor into modules and extract styles
