+++ index.html
@@ -8,77 +8,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="css/style.css">
-    <style>
-        /* Retro video player in bottom-right corner */
-        .retro-player {
-            position: fixed;
-            left: 0;
-            bottom: 410px; /* 10px gap above chat (400px tall) */
-            width: 320px;            /* base width */
-            aspect-ratio: 4 / 3;     /* classic TV commercials ratio */
-            height: auto;            /* auto-calculated from aspect ratio */
-            background: #000;
-            border: 4px solid #666666;
-            box-shadow: 4px 4px 0px #333333;
-            z-index: 30;
-            overflow: hidden; /* ensure cover stays inside */
-            transition: width 0.3s, bottom 0.3s;
-        }
-
-        .retro-player.compressed {
-            width: 200px;           /* match compressed chat width */
-            bottom: 260px;          /* 10px gap above 250px compressed chat */
-        }
-
-        .retro-player iframe {
-            width: 100%;
-            height: 100%;
-            border: none;
-            pointer-events: none; /* disable hover UI & controls */
-        }
-
-        /* Native video element styling */
-        .retro-player video {
-            width: 100%;
-            height: 100%;
-            border: none;
-            pointer-events: none;
-            object-fit: cover;
-        }
-
-        /* Black screen cover that fades out */
-        .retro-player .video-cover {
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            background: transparent;
-            z-index: 2;
-            pointer-events: auto; /* allow click to toggle mute */
-            animation: none;
-        }
-
-        /* Mute icon (speaker with slash) */
-        .mute-icon {
-            position: absolute;
-            bottom: 6px;
-            right: 6px;
-            width: 24px;
-            height: 24px;
-            background: rgba(0,0,0,0.6);
-            border-radius: 4px;
-            cursor: pointer;
-            z-index: 3;
-            display: none; /* only visible when muted */
-            padding: 2px;
-        }
-        .mute-icon img {
-            width: 100%;
-            height: 100%;
-            filter: invert(1); /* make lines white against dark bg */
-        }
-    </style>
+    <link rel="stylesheet" href="css/retro-player.css">
 </head>
 <body>
     <button class="login-btn">LOG IN</button>
@@ -291,7 +221,7 @@
         </div>
     </div>
 
-    <script src="js/script.js"></script>
+    <script type="module" src="js/main.js"></script>
     <script>
         // ******** Retro-ad player logic using Google Drive videos ********
         document.addEventListener('DOMContentLoaded', () => {
