+++ js/config.js
@@ -1,0 +1,23 @@
+export const FIELD = {
+  MARGIN: 20,
+  GOAL_HEIGHT: 120
+};
+
+export const CAR = {
+  ACCELERATION: 0.1,
+  MAX_SPEED: 4,
+  FRICTION: 0.03,
+  TURN_SPEED: 0.05,
+  FORWARD_FRICTION: 0.02,
+  SIDE_FRICTION_NORMAL: 0.3,
+  SIDE_FRICTION_BRAKE: 0.03
+};
+
+export const BALL = {
+  FRICTION: 0.98
+};
+
+export const COLORS = {
+  PLAYER1: '#c62828',
+  PLAYER2: '#2962ff'
+};
